chore: fix dependency cycle reference
